BitmapFontFactory (removed) => BitmapFont.from() API created.
+ Support for character ranges (e.g. chars: [['a', 'z'], ['A', 'Z']]
+ Support for arrays of characters (e.g. chars: ['abcdef', 'GHIJK'])
+ Documented drawGlyph (local, internal function)
+ Mandatory arguments for BitmapFont.from()
