+++ packages/text-bitmap/src/index.ts
@@ -2,4 +2,3 @@ export * from './BitmapText';
 export * from './BitmapFontLoader';
 export * from './BitmapFont';
 export * from './BitmapFontData';
-export * from './BitmapFontFactory';
